why us 부분 개발 완료
Result :  (('Weak', 'Your essay seems to be lacking some details about the college and may not demonstrate a strong interest. You may consider doing more research on the college and department you wish to study in.'), ('Weak', "Regarding your fit with the intended major, your knowledge of the discipline's intellectual concepts seems lacking."), 'What you do for fun essay is an excellent opportunity to share something about you without bragging about an award or a big leadership position. Experts say that it is better to be down-to-earth with this type of prompt and write about your favorite pastime, even if it may sound trivial.', (2, ['excitement', 'admiration'], 33.33, 'Mediocre'), 'Mediocre')